+++ textrepo-app/src/main/java/nl/knaw/huc/api/ResultIndexer.java
@@ -8,11 +8,15 @@ public class ResultIndexer {
   private final String name;
   private final List<String> hosts;
   private final List<String> mimetypes;
+  private final ResultIndexerFieldsConfiguration fields;
+  private final String mapping;
 
   public ResultIndexer(MappedIndexerConfiguration indexer) {
     this.name = indexer.name;
     this.hosts = indexer.elasticsearch.hosts;
     this.mimetypes = indexer.mimetypes;
+    this.fields = new ResultIndexerFieldsConfiguration(indexer.fields);
+    this.mapping = indexer.mapping;
   }
 
   public String getName() {
@@ -25,5 +29,13 @@ public class ResultIndexer {
 
   public List<String> getMimetypes() {
     return mimetypes;
+  }
+
+  public ResultIndexerFieldsConfiguration getFields() {
+    return fields;
+  }
+
+  public String getMapping() {
+    return mapping;
   }
 }