+++ textrepo-app/src/main/java/nl/knaw/huc/resources/rest/VersionContentsResource.java
@@ -7,6 +7,7 @@ import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import nl.knaw.huc.api.ResultVersion;
 import nl.knaw.huc.exceptions.MethodNotAllowedException;
+import nl.knaw.huc.helpers.ContentsHelper;
 import nl.knaw.huc.service.version.content.VersionContentsService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,6 +16,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -23,8 +25,8 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
 import java.util.UUID;
 
+import static javax.ws.rs.core.HttpHeaders.ACCEPT_ENCODING;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
-import static nl.knaw.huc.helpers.ContentsHelper.getContentsAsAttachment;
 
 @Api(tags = {"versions", "contents"})
 @Path("/rest/versions/{versionId}/contents")
@@ -36,12 +38,11 @@ public class VersionContentsResource {
   private static final String DELETE_ERROR_MSG = "Not allowed to delete contents of version: delete version instead";
 
   private final VersionContentsService contentsService;
-  private final int decompressLimit;
+  private final ContentsHelper contentsHelper;
 
-  public VersionContentsResource(VersionContentsService contentsService, int contentDecompressionLimit) {
+  public VersionContentsResource(VersionContentsService contentsService, ContentsHelper contentsHelper) {
     this.contentsService = contentsService;
-    this.decompressLimit = contentDecompressionLimit;
-    log.debug("contentDecompressionLimit={}", decompressLimit);
+    this.contentsHelper = contentsHelper;
   }
 
   @POST
@@ -58,13 +59,14 @@ public class VersionContentsResource {
   @ApiOperation(value = "Retrieve version contents")
   @ApiResponses(value = {@ApiResponse(code = 200, response = ResultVersion.class, message = "OK")})
   public Response get(
+      @HeaderParam(ACCEPT_ENCODING) String acceptEncoding,
       @PathParam("versionId") @NotNull @Valid UUID versionId
   ) {
     log.debug("Get version contents: versionId={}", versionId);
     var contents = contentsService.getByVersionId(versionId);
     log.debug("Got version contents: {}", contents);
 
-    return getContentsAsAttachment(contents, decompressLimit).build();
+    return contentsHelper.asAttachment(contents, acceptEncoding).build();
   }
 
   @PUT