+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- File indexer: for more info, see readme: `./elasticsearch/file`
+- Origin Link header to indexer fields request
 - Show 'about' page `index.html` (with link to `index.json`) on nginx landing page, with version info and links to services, documentation and source code.
+- Show version and configuration info at root of textrepo API
 - Resources now send 'Content-Encoding: gzip' when sending gzip'd contents
 
 ### Changed
@@ -15,6 +18,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  (as it is named in docker builds: https://docs.docker.com/docker-hub/builds/advanced)
 - Renamed env var `NGINX_ROOT_REDIRECT` to `CONCORDION_ROOT_REDIRECT`
 - Database queries using 'count' now use 'long' as their Java counterpart (was: int)
+
+## [v1.10]
+- [TT-631] When yielding contents, return gzip'd content when 'Accept-Encoding: gzip'
+- [TT-645] Resources now send 'Content-Encoding: gzip' when sending gzip'd contents
+- [TT-574] Database queries using 'count' now use 'long' as their Java counterpart (was: int)
 
 ### Fixed
 - Use `tar.gz` instead of `tar` as backup extension